DrMZ

DrMZ

DrMZ.advection_diffusion_pde! — Method

advection_diffusion_pde!(duhat,uhat,p,t)

RHS for the advection-diffusion equation $u_t = - u_x + ν u_{xx}$ for numerical integration in Fourier space where ν is the viscosity.

DrMZ.advection_pde! — Method

advection_pde!(duhat,uhat,p,t)

RHS for the advection equation $u_t = - u_x$ for numerical integration in Fourier space.

DrMZ.average_error — Method

average_error(domain,error)

Compute the average error using the trapezoid rule $\frac{1}{T} \int_0^T error(t) dt$.

DrMZ.average_ic_error — Method

average_ic_error(target,prediction)

Compute the two-norm relative error between the prediction and target values for an initial condition.

DrMZ.basis_derivative — Method

basis_derivative(basis,nodes)

Compute the derivatives of each basis function using auto differentiation.

DrMZ.basis_eval — Method

basis_eval(basis,nodes)

Evaluate the basis functions for a specified spatial grid or at a specified point. Evaluation for a specified grid returns an $N x M$ matrix with $N$ respresenting the number of nodes and $M$ representing the number of custom basis functions. Evalaution for a specified point returns a $1 x M$ vector.

DrMZ.basis_interpolate — Method

basis_interpolate(coeffs,x,L1,L2)

DrMZ.build_basis — Method

build_basis(trunk,L1,L2,M,nodes,weights,L;cutoff=1e-13,tol=1e-12)

Generate a dictionary of orthonormal custom basis functions based on a specified quadrature rule. An orthogonal polynomial expansion is utilized for the evaluation at points away from the specified quadrature nodes.

DrMZ.build_basis_expand — Method

build_basis_expand(trunk,L1,L2,M,nodes,weights;cutoff=1e-13,tol=1e-12)

Generate a dictionary of periodic, orthonormal custom basis functions based on the trapezoid rule. B-splines are utilized for the evaluation at points away from the specified quadrature nodes.

To Do: Add time-sampling capabilities...

DrMZ.build_dense_model — Method

function build_dense_model(number_layers,neurons,activations)

Build a feedforward neural network (FFNN) consisting of number_layers of Flux dense layers for the specified number of neurons and activations.

Examples

julia> build_dense_model(2,[(128,128),(128,128)],[relu,relu])
Chain(Dense(128, 128, NNlib.relu), Dense(128, 128, NNlib.relu))

DrMZ.burgers_flux — Method

burgers_flux(u_right,u_left)

Numerical flux for Burgers nonlinearity, $f(u) = \frac{1}{2}u^2$, for use with discontinuous Galerkin method.

DrMZ.central_difference — Method

central_difference(u_j,u_jpos,u_jneg,mu)

Compute the second order central difference for the viscous term of the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$. mu is equal to $ \frac{\nu}{\Delta x^2}$.

DrMZ.cheby_diff — Method

cheby_diff(sol,N,L1,L2)

Compute the derivative of using a Chebyshev differentiation matrix on the interval $[L1,L2]$ with N discretization points.

DrMZ.cheby_diff_matrix — Method

cheby_diff_matrix(N,L1,L2)

Generate the Chebyshev differentiation matrix for the interval $[L1,L2]$ with N discretization points.

DrMZ.cheby_grid — Method

cheby_grid(N,L1,L2)

Generate the grid of Chebyshev points on the interval $[L1,L2]$ with N discretization points.

DrMZ.clenshaw_curtis — Method

clenshaw_curtis(N,L1,L2)

Compute the nodes and weights for Clenshaw-Curtis quadrature with N discretization points and on the interval $[L1,L2]$.

DrMZ.dbasis_interpolate — Method

dbasis_interpolate(coeffs,x,L1,L2,m)

DrMZ.dlegendre_norm — Method

dlegendre_norm(x,L1,L2,n)

Compute the 1st derivative of the n-th shifted Legendre polynomial

DrMZ.dlegendre_norm_basis_build — Method

dlegendre_norm_basis_build(nmax,L1,L2)

DrMZ.error_rel — Method

error_rel(target,prediction)

Compute the relative error between the prediction and target values.

DrMZ.error_se — Method

error_se(target,prediction)

Compute the squared error between the prediction and target values.

DrMZ.exp_kernel_periodic — Method

function exp_kernel_periodic(fnc,x_locations;length_scale=0.5)

Covariance kernel for radial basis function (GRF) and IC function for domain.

DrMZ.expansion_approximation — Method

expansion_approximation(basis,coefficients,nodes)

Compute the expansion approximation...

DrMZ.expansion_coefficients — Method

expansion_coefficients(basis,fnc,nodes,weights)

Compute the expansion coefficients...

DrMZ.feature_expansion_set — Method

feature_expansion_set(L1,L2,neurons,solution_loc)

DrMZ.feature_expansion_set_x — Method

feature_expansion_set_x(L1,L2,neurons,x_loc)

DrMZ.feature_expansion_single — Method

feature_expansion_single(L1,L2,neurons,x)

DrMZ.fft_norm — Method

fft_norm(solution)

Compute the FFT normalized by $\frac{1}{N}$.

DrMZ.fl — Method

fl(ulv,urv,ubv)

DrMZ.fourier_diff — Method

fourier_diff(sol,N,dL;format="matrix")

Compute the derivative using a Fourier differentiation matrix (default) or the spectral derivative for periodic functions for domain length dL and with N discretization points.

DrMZ.gauss_legendre — Method

gauss_legendre(N,L1,L2)

Compute the nodes and weights for Gauss-Legendre quadrature with N discretization points and on the interval $[L1,L2]$.

DrMZ.generate_basis_solution — Method

generate_basis_solution(nodes,weights,tspan,initial_condition,basis,params,pde_function;dt=1e-3,rtol=1e-10,atol=1e-14,peak=1.05)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a custom basis function expansion and a RK45 solver.

DrMZ.generate_basis_solution_esdirk — Method

generate_basis_solution_esdirk(nodes,weights,tspan,initial_condition,basis,params,pde_function;dt=1e-4,rtol=1e-8,atol=1e-12)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a custom basis function expansion and a ESDIRK4 solver.

DrMZ.generate_basis_solution_implicit — Method

generate_basis_solution_implicit(nodes,weights,tspan,initial_condition,basis,params,pde_function;dt=1e-3,rtol=1e-8,atol=1e-12)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a custom basis function expansion and a Crank-Nicolson solver.

DrMZ.generate_fourier_solution — Method

generate_fourier_solution(L1,L2,tspan,N,initial_condition,pde_function;dt=1e-3,nu=0.1,rtol=1e-10,atol=1e-14)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a N mode Fourier expansion and a RK45 solver.

DrMZ.generate_fourier_solution_esdirk — Method

generate_fourier_solution_esdirk(L1,L2,tspan,N,initial_condition,pde_function;dt=1e-4,rtol=1e-8,atol=1e-12,nu=0.1)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a N mode Fourier expansion and a ESDIRK4 solver.

DrMZ.generate_fourier_solution_implicit — Method

generate_fourier_solution_implicit(L1,L2,tspan,N,initial_condition,pde_function_explicit,pde_function_implicit;dt=1e-4,rtol=1e-8,atol=1e-12,nu=0.1)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a N mode Fourier expansion and a Crank-Nicolson solver.

DrMZ.generate_muscl_minmod_solution — Method

generate_muscl_minmod_solution(L1,L2,t_end,N,u0,pde_function_handle;dt=1e-4,kappa=-1)

DrMZ.generate_periodic_functions — Method

generate_periodic_functions(fnc,x_locations,number_functions,length_scale)

Generate a specified number_functions of random periodic vectors using the exp_kernel_periodic function and a multivariate distribution.

DrMZ.generate_periodic_train_test — Method

generate_periodic_train_test(t_span,number_sensors,number_train_functions,number_test_functions,number_solution_points,pde_function_handle;L1=0,L2=2*pi,length_scale=0.5,batch=number_solution_points,dt=1e-3,nu_val=0.1,fnc=(x)->sin(x/2)^2)

Generate the training and testing data for a specified pde_function_handle for periodic boundary conditions using a Fourier spectral method. Defaults to IC $f(\sin^2(x/2))$ and $x ∈ [0,1]$.

DrMZ.generate_periodic_train_test_esdirk — Method

generate_periodic_train_test_esdirk(t_span,number_sensors,number_train_functions,number_test_functions,number_solution_points,pde_function_handle;L1=0,L2=2*pi,length_scale=0.5,batch=number_solution_points,dt_size=1e-4,nu_val=0.1,domain="periodic",fnc=(x)->sin(x/2)^2,mode_multiplier=4)

Generate the training and testing data for a specified pde_function_handle for periodic boundary conditions using a Fourier spectral method and a ESDIRK ODE solver. Defaults to IC $f(\sin^2(x/2))$ and $x ∈ [0,1]$.

DrMZ.generate_periodic_train_test_implicit — Method

generate_periodic_train_test_implicit(t_span,number_sensors,number_train_functions,number_test_functions,number_solution_points,pde_function_handle;L1=0,L2=2*pi,length_scale=0.5,batch=number_solution_points,dt_size=1e-4,nu_val=0.1,domain="periodic",fnc=(x)->sin(x/2)^2,mode_multiplier=4)

Generate the training and testing data for a specified pde_function_handle for periodic boundary conditions using a Fourier spectral method and a Crank-Nicolson solver. Defaults to IC $f(\sin^2(x/2))$ and $x ∈ [0,1]$.

DrMZ.generate_periodic_train_test_initial_conditions — Method

generate_periodic_train_test_initial_conditions(t_span,number_sensors,number_test_functions,number_train_functions,number_solution_points,pde_function_handle;L1=0,L2=2*pi,length_scale=0.5,batch=number_solution_points,dt=1e-3,nu_val=0.1,domain="periodic",fnc=(x)->sin(x/2)^2)

Generate the training and testing data for a specified pde_function_handle for periodic boundary conditions using a Fourier spectral method. Defaults to IC $f(\sin^2(x/2))$ and $x ∈ [0,1]$.

DrMZ.generate_periodic_train_test_muscl — Method

generate_periodic_train_test_muscl(t_span,number_sensors,number_test_functions,number_train_functions,number_solution_points,pde_function_handle;L1=0,L2=2*pi,length_scale=0.5,batch=number_solution_points,dt_size=1e-4,upwind_solution_points=4096,fnc=(x)->sin(x/2)^2)

Generate the training and testing data for a specified pde_function_handle for periodic boundary conditions using a MUSCL method. Defaults to IC $f(\sin^2(x/2))$ and $x ∈ [0,1]$.

DrMZ.generate_sinusoidal_functions_2_parameter — Method

generate_sinusoidal_functions_2_parameter(x_locations,number_functions)

Generate a specified number_functions of random periodic vectors for the distribution $α \sin(x)+β$ for $α ∈ [-1,1]$ and $β ∈ [-1,1]$.

DrMZ.get_1D_energy_custom — Method

get_1D_energy_custom(basis,u_solution,L1,L2,weights;multiplier=1/(4*pi))

Compute the energy in the custom basis domain: $ \frac{1}{2} \sum \vert \a_k \vert^2 $. Multiplier defaults to $ \frac{1}{4\pi} $ to match the Fourier calculation.

DrMZ.get_1D_energy_custom_coefficients — Method

get_1D_energy_custom_coefficients(u_coefficients;zeta=1/(4*pi))

Compute the energy in the custom basis domain: $ \zeta \sum \vert \a_k \vert^2 $. $ \zeta $ defaults to $ \frac{1}{4\pi} $ to match the Fourier calculation.

DrMZ.get_1D_energy_fft — Method

get_1D_energy_fft(u_solution)

Compute the energy in the Fourier domain using the scaling of $ \frac{1}{N} $. Note: this does not include the 2π multiplier found in Parseval's identity for Fourier series and computes $ \frac{1}{2} \sum \vert \hat{u}_k \vert^2 $.

DrMZ.get_1D_energy_upwind — Method

get_1D_energy_upwind(u_solution_full,u_solution,N)

DrMZ.ic_error — Method

ic_error(target,prediction)

Compute the relative error between the prediction and target values for an initial condition. If the target = 0, the target in denomenator is augmented by $\epsilon_{machine}$.

DrMZ.ifft_norm — Method

ifft_norm(solution)

Compute the IFFT normalized by $N$.

DrMZ.inviscid_burgers_pde! — Method

inviscid_burgers_pde!(duhat,uhat,p,t)

RHS for the inviscid Burgers equation $u_t = - u u_x$ for numerical integration in Fourier space.

DrMZ.kdv_explicit_pde! — Method

kdv_explicit_pde!(duhat,uhat,p,t)

Explicit portion of the RHS for the Korteweg-de Vries equation $u_t = - u u_x - \nu^2 u_{xxx}$ for numerical integration numerical integration in Fourier space.

DrMZ.kdv_implicit_pde! — Method

kdv_implicit_pde!(duhat,uhat,p,t)

Implicit portion of the RHS for the Korteweg-de Vries equation $u_t = - u u_x - \nu^2 u_{xxx}$ for numerical integration numerical integration in Fourier space.

DrMZ.kdv_pde! — Method

kdv_pde!(duhat,uhat,p,t)

RHS for the Korteweg-de Vries equation $u_t = - u u_x - \nu^2 u_{xxx}$ for numerical integration numerical integration in Fourier space.

DrMZ.ks_explicit_pde! — Method

ks_explicit_pde!(duhat,uhat,p,t)

Explicit portion of the RHS for the Kuramoto-Sivashinsky equation $u_t = - u u_x -u_{xx} - \nu u_{xxxx}$ for numerical integration numerical integration in Fourier space.

DrMZ.ks_implicit_pde! — Method

ks_implicit_pde!(duhat,uhat,p,t)

Implicit portion of the RHS for the Kuramoto-Sivashinsky equation $u_t = - u u_x -u_{xx} - \nu u_{xxxx}$ for numerical integration numerical integration in Fourier space.

DrMZ.ks_pde! — Method

ks_pde!(duhat,uhat,p,t)

RHS for the Kuramoto-Sivashinsky equation $u_t = - u u_x -u_{xx} - \nu u_{xxxx}$ for numerical integration numerical integration in Fourier space.

DrMZ.legendre_norm — Method

legendre_norm(x,L1,L2,n)

Compute the n-th orthonormal, shifted Legendre polynomial

DrMZ.legendre_norm_basis_build — Method

legendre_norm_basis_build(nmax,L1,L2)

DrMZ.linear_reg — Method

linear_reg(x,y)

DrMZ.load_basis — Method

load_basis(basis,M,pde_function)

DrMZ.load_branch — Method

load_branch(n_epoch,pde_function)

Load the trained branch neural networks.

DrMZ.load_data — Method

load_data(n_epoch,number_train_functions,number_test_functions,pde_function)

Load the trained branch and trunk neural networks along with the train_data and test_data.

DrMZ.load_data_initial_conditions — Method

load_data_initial_conditions(number_train_functions,number_test_functions,pde_function)

Load the initial conditions from the train_data and test_data.

DrMZ.load_data_train_test — Method

load_data_train_test(number_train_functions,number_test_functions,pde_function)

Load the train_data and test_data.

DrMZ.load_model — Method

load_model(n_epoch,pde_function)

Load the trained branch and trunk neural networks.

DrMZ.load_trunk — Method

load_trunk(n_epoch,pde_function)

Load the trained trunk neural networks.

DrMZ.loss_all — Method

loss_all(branch,trunk,initial_conditon,solution_location,target_value)

Compute the mean squared error (MSE) for a complete dataset.

DrMZ.minmod — Method

minmod(x,y)

DrMZ.mode_extractor — Method

mode_extractor(uhat,N)

DrMZ.mse_error — Method

mse_error(target,prediction)

Compute the mean squared error between the prediction and target values.

DrMZ.muscl_minmod_RHS! — Method

muscl_minmod_RHS!(du,u,p,t)

DrMZ.muscl_minmod_viscous_RHS! — Method

muscl_minmod_viscous_RHS!(du,u,p,t)

DrMZ.norm_infinity_error — Method

norm_infinity_error(target,prediction)

Compute the two-norm relative error between the prediction and target values.

DrMZ.norm_rel_error — Method

norm_rel_error(target,prediction)

Compute the Euclidean two-norm relative error between the prediction and target values.

DrMZ.norm_rel_error_continuous — Method

function norm_rel_error_continuous(target,prediction,weights)

Compute the continuous two-norm relative error between the prediction and target values.

DrMZ.orthonormal_check — Method

orthonormal_check(basis,weights;tol = 1e-12)

DrMZ.periodic_fill_domain — Method

periodic_fill_domain(x_locations)

Output the full domain from periodic domain specified for x_locations.

DrMZ.periodic_fill_solution — Method

periodic_fill_solution(solution)

Output the full $u(t,x)$ solution from periodic solution.

DrMZ.predict — Method

predict(branch,trunk,initial_condition,x_locations,t_values)

Predict solution $u(t,x)$ at specified output locations using trained operator neural network branch and trunk.

DrMZ.quadratic_nonlinear — Method

quadratic_nonlinear!(uhat,N,dL,alpha)

Compute the convolution sum $\frac{ik}{2}\sum_{p+q=k} u_p u_q$ resulting from the quadratic nonlinearity of Burgers equation $u u_x$ in Fourier space. Convolution sum is padded with the 3/2 rule for dealiasing.

DrMZ.quadratic_nonlinear_basis — Method

quadratic_nonlinear_basis(u,nonlinear_triple)

DrMZ.quadratic_nonlinear_triple_product_basis — Method

quadratic_nonlinear_triple_product_basis(basis_nodes,Dbasis_nodes,nodes,weights)

DrMZ.quadratic_nonlinear_triple_product_basis_galerkin — Method

quadratic_nonlinear_triple_product_basis_galerkin(basis_nodes,Dbasis_nodes,nodes,weights)

DrMZ.reduced_initial_condition — Method

reduced_initial_condition(L1,L2,x_reduced,x_locations,initial_condition)

Extract the $x$ locations and intial condition values $u(x)$ at a reduced number of equally spaced spatial locations.

DrMZ.rhs_advection! — Method

rhs_advection!(du,u,p,t)

RHS for the advection equation $u_t = - u_x$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_advection_diffusion! — Method

rhs_advection_diffusion!(du,u,p,t)

RHS for the advection-diffusion equation $u_t = - u_x + ν u_{xx}$ for numerical integration in custom basis space using discontinuous Galerkin method and where ν is the viscosity.

DrMZ.rhs_advection_diffusion_dirichlet! — Method

rhs_advection_diffusion_dirichlet!(du,u,p,t)

RHS for the advection-diffusion equation $u_t = - u_x + ν u_{xx}$ with Dirichlet boundary conditions for numerical integration in custom basis space using discontinuous Galerkin method and where ν is the viscosity.

DrMZ.rhs_advection_diffusion_galerkin! — Method

rhs_advection_diffusion_galerkin!(du,u,p,t)

DrMZ.rhs_advection_galerkin! — Method

rhs_advection_galerkin!(du,u,p,t)

DrMZ.rhs_explicit_kdv! — Method

rhs_explicit_kdv!(du,u,p,t)

Explicit portion of the RHS for the Korteweg-de Vries equation $u_t = - u u_x - \nu^2 u_{xxx}$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_explicit_ks! — Method

rhs_explicit_ks!(du,u,p,t)

Explicit portion of the RHS for the Kuramoto-Sivashinsky equation $u_t = - u u_x -u_{xx} - \nu u_{xxxx}$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_implicit_kdv! — Method

rhs_implicit_kdv!(du,u,p,t)

Implicit portion of the RHS for the Korteweg-de Vries equation $u_t = - u u_x - \nu^2 u_{xxx}$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_implicit_ks! — Method

rhs_implicit_ks!(du,u,p,t)

Implicit portion of the RHS for the Kuramoto-Sivashinsky equation $u_t = - u u_x -u_{xx} - \nu u_{xxxx}$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_inviscid_burgers! — Method

rhs_inviscid_burgers!(du,u,p,t)

RHS for the inviscid Burgers equation $u_t = - u u_x$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_inviscid_burgers_galerkin! — Method

rhs_inviscid_burgers_galerkin!(du,u,p,t)

DrMZ.rhs_inviscid_burgers_pseudo! — Method

rhs_inviscid_burgers_pseudo!(du,u,p,t)

DrMZ.rhs_kdv! — Method

rhs_kdv!(du,u,p,t)

RHS for the Korteweg-de Vries equation $u_t = - u u_x - \nu^2 u_{xxx}$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_kdv_galerkin! — Method

rhs_kdv_galerkin!(du,u,p,t)

DrMZ.rhs_kdv_pseudo! — Method

rhs_kdv_pseudo!(du,u,p,t)

DrMZ.rhs_ks! — Method

rhs_ks!(du,u,p,t)

RHS for the Kuramoto-Sivashinsky equation $u_t = - u u_x -u_{xx} - \nu u_{xxxx}$ for numerical integration in custom basis space using discontinuous Galerkin method.

DrMZ.rhs_ks_galerkin! — Method

rhs_ks_galerkin!(du,u,p,t)

DrMZ.rhs_ks_pseudo! — Method

rhs_ks_pseudo!(du,u,p,t)

No padding to account for the impact of potential aliasing

DrMZ.rhs_viscous_burgers! — Method

rhs_viscous_burgers!(du,u,p,t)

RHS for the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$ for numerical integration in custom basis space using Discontinuous Galerking method and where ν is the viscosity.

DrMZ.rhs_viscous_burgers_galerkin! — Method

rhs_viscous_burgers_galerkin!(du,u,p,t)

To Do: Simplify input parameters - typical all Galerkin RHS's

DrMZ.rhs_viscous_burgers_pseudo! — Method

rhs_viscous_burgers_pseudo!(du,u,p,t)

DrMZ.save_basis — Method

save_basis(basis,M,pde_function)

DrMZ.save_data — Method

save_data(train_data,test_data,number_train_functions,number_test_functions,number_solution_points,pde_function)

Save the train_data and test_data.

DrMZ.save_data_initial_conditions — Method

save_data_initial_conditions(number_train_functions,number_test_functions)

Saves the initial conditions from the train_ic and test_ic.

DrMZ.save_model — Method

save_model(branch,trunk,n_epoch,loss_all_train,loss_all_test,pde_function)

Save the trained branch and trunk neural networks and the training and testing loss history.

DrMZ.solution_extraction — Method

solution_extraction(x_locations,t_values,solution,initial_condition,number_solution_points)

Extract the specified number_solution_points randomly from the $u(t,x)$ solution space.

DrMZ.solution_interpolation — Method

solution_interpolation(t_span_original,x_locations_original,t_span_interpolate,x_locations_interpolate,solution)

Compute the $u(t,x)$ solution at intermediate $(t,x)$ locations using linear interpolation.

DrMZ.solution_spatial_sampling — Method

solution_spatial_sampling(x_prediction,x_target,solution)

Extract the solution values $u(t,x)$ at a reduced number of equally spaced spatial locations.

DrMZ.solution_temporal_sampling — Method

solution_temporal_sampling(t_prediction,t_target,solution)

Extract the solution values $u(t,x)$ at a reduced number of equally spaced temporal locations.

DrMZ.spectral_approximation_fourier — Method

spectral_approximation_fourier(x_locations,k,coefficients)

DrMZ.train_model — Method

train_model(branch,trunk,n_epoch,train_data;learning_rate=0.00001,save_at=2500,starting_epoch=0)

Train the operator neural network using the mean squared error (MSE) and Adam optimization for n_epochs epochs.

DrMZ.trapezoid — Method

trapezoid(N,L1,L2)

Compute the nodes and weights for trapezoid rule with N discretization points and on the interval $[0,L2]$. TO DO: Revise for [L1,L2] support.

DrMZ.trapz — Method

trapz(x_range,integrand)

Numerical integration using the multi-application trapezoidal rule.

DrMZ.trunk_build — Method

trunk_build(trunk,M,dtsample)

Generate a dictionary of trunk functions evaluated at $t = 0$ or at a range of times spanning (0:dtsample:1)

DrMZ.trunk_ortho_build — Method

trunk_ortho_build(utilde,L1,L2,nodes,weights,L)

Generate a dictionary of orthonormal custom basis functions which use an orthogonal polynomial expansion for evaluation away from the specified quadrature nodes.

DrMZ.trunk_ortho_build_expand — Method

trunk_ortho_build_expand(utilde,nodes;p=17)

Generate a dictionary of periodic, orthonormal custom basis functions which use B-splines for evaluation away from the specified quadrature nodes.

DrMZ.ub — Method

ub(ulv,urv)

DrMZ.ulnl — Method

ulnl(uj,ujn,ujnn,kappa,omega)

DrMZ.ulpl — Method

ulpl(ujp,uj,ujn,kappa,omega)

DrMZ.urnl — Method

urnl(ujp,uj,ujn,kappa,omega)

DrMZ.urpl — Method

urpl(ujpp,ujp,uj,kappa,omega)

DrMZ.viscous_burgers_pde! — Method

viscous_burgers_pde!(duhat,uhat,p,t)

RHS for the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$ for numerical integration in Fourier space where ν is the viscosity.

DrMZ.advection_diffusion_pde!

DrMZ.advection_pde!

DrMZ.average_error

DrMZ.average_ic_error

DrMZ.basis_derivative

DrMZ.basis_eval

DrMZ.basis_interpolate

DrMZ.build_basis

DrMZ.build_basis_expand

DrMZ.build_dense_model

DrMZ.burgers_flux

DrMZ.central_difference

DrMZ.cheby_diff

DrMZ.cheby_diff_matrix

DrMZ.cheby_grid

DrMZ.clenshaw_curtis

DrMZ.dbasis_interpolate

DrMZ.dlegendre_norm

DrMZ.dlegendre_norm_basis_build

DrMZ.error_rel

DrMZ.error_se

DrMZ.exp_kernel_periodic

DrMZ.expansion_approximation

DrMZ.expansion_coefficients

DrMZ.feature_expansion_set

DrMZ.feature_expansion_set_x

DrMZ.feature_expansion_single

DrMZ.fft_norm

DrMZ.fl

DrMZ.fourier_diff

DrMZ.gauss_legendre

DrMZ.generate_basis_solution

DrMZ.generate_basis_solution_esdirk

DrMZ.generate_basis_solution_implicit

DrMZ.generate_fourier_solution

DrMZ.generate_fourier_solution_esdirk

DrMZ.generate_fourier_solution_implicit

DrMZ.generate_muscl_minmod_solution

DrMZ.generate_periodic_functions

DrMZ.generate_periodic_train_test

DrMZ.generate_periodic_train_test_esdirk

DrMZ.generate_periodic_train_test_implicit

DrMZ.generate_periodic_train_test_initial_conditions

DrMZ.generate_periodic_train_test_muscl

DrMZ.generate_sinusoidal_functions_2_parameter

DrMZ.get_1D_energy_custom

DrMZ.get_1D_energy_custom_coefficients

DrMZ.get_1D_energy_fft

DrMZ.get_1D_energy_upwind

DrMZ.ic_error

DrMZ.ifft_norm

DrMZ.inviscid_burgers_pde!

DrMZ.kdv_explicit_pde!

DrMZ.kdv_implicit_pde!

DrMZ.kdv_pde!

DrMZ.ks_explicit_pde!

DrMZ.ks_implicit_pde!

DrMZ.ks_pde!

DrMZ.legendre_norm

DrMZ.legendre_norm_basis_build

DrMZ.linear_reg

DrMZ.load_basis

DrMZ.load_branch

DrMZ.load_data

DrMZ.load_data_initial_conditions

DrMZ.load_data_train_test

DrMZ.load_model

DrMZ.load_trunk

DrMZ.loss_all

DrMZ.minmod

DrMZ.mode_extractor

DrMZ.mse_error

DrMZ.muscl_minmod_RHS!

DrMZ.muscl_minmod_viscous_RHS!

DrMZ.norm_infinity_error

DrMZ.norm_rel_error

DrMZ.norm_rel_error_continuous

DrMZ.orthonormal_check

DrMZ.periodic_fill_domain

DrMZ.periodic_fill_solution

DrMZ.predict

DrMZ.quadratic_nonlinear

DrMZ.quadratic_nonlinear_basis

DrMZ.quadratic_nonlinear_triple_product_basis

DrMZ.quadratic_nonlinear_triple_product_basis_galerkin

DrMZ.reduced_initial_condition

DrMZ.rhs_advection!

DrMZ.rhs_advection_diffusion!

DrMZ.rhs_advection_diffusion_dirichlet!

DrMZ.rhs_advection_diffusion_galerkin!

DrMZ.rhs_advection_galerkin!

DrMZ.rhs_explicit_kdv!

DrMZ.rhs_explicit_ks!

DrMZ.rhs_implicit_kdv!

DrMZ.rhs_implicit_ks!

DrMZ.rhs_inviscid_burgers!

DrMZ.rhs_inviscid_burgers_galerkin!

DrMZ.rhs_inviscid_burgers_pseudo!

DrMZ.rhs_kdv!

DrMZ.rhs_kdv_galerkin!

DrMZ.rhs_kdv_pseudo!

DrMZ.rhs_ks!

DrMZ.rhs_ks_galerkin!

DrMZ.rhs_ks_pseudo!

DrMZ.rhs_viscous_burgers!

DrMZ.rhs_viscous_burgers_galerkin!

DrMZ.rhs_viscous_burgers_pseudo!

DrMZ.save_basis

DrMZ.save_data

DrMZ.save_data_initial_conditions

DrMZ.save_model

DrMZ.solution_extraction

DrMZ.solution_interpolation

DrMZ.solution_spatial_sampling

DrMZ.solution_temporal_sampling

DrMZ.spectral_approximation_fourier

DrMZ.train_model

DrMZ.trapezoid

DrMZ.trapz

DrMZ.trunk_build

DrMZ.trunk_ortho_build

DrMZ.trunk_ortho_build_expand

DrMZ.ub

DrMZ.ulnl

DrMZ.ulpl

DrMZ.urnl

DrMZ.urpl

DrMZ.viscous_burgers_pde!