DrMZ

DrMZ.advection_diffusion_pde!

DrMZ.advection_pde!

DrMZ.average_error

DrMZ.average_ic_error

DrMZ.backward_upwind

DrMZ.backward_upwind_limited

DrMZ.basis_OpNN

DrMZ.build_basis

DrMZ.build_basis_redefinition

DrMZ.build_dense_model

DrMZ.cheby_diff

DrMZ.cheby_diff_matrix

DrMZ.cheby_grid

DrMZ.error_test_rel

DrMZ.error_test_sse

DrMZ.exp_kernel_periodic

DrMZ.fft_norm

DrMZ.forward_upwind

DrMZ.forward_upwind_limited

DrMZ.fourier_diff

DrMZ.generate_basis_solution

DrMZ.generate_basis_solution_nonlinear

DrMZ.generate_bwlimitersoupwind_solution

DrMZ.generate_fourier_solution

DrMZ.generate_fourier_solution_split

DrMZ.generate_periodic_functions

DrMZ.generate_periodic_train_test

DrMZ.generate_sinusoidal_functions_2_parameter

DrMZ.gradient_ratio_backward_j

DrMZ.gradient_ratio_backward_jneg

DrMZ.gradient_ratio_forward_j

DrMZ.gradient_ratio_forward_jpos

DrMZ.gram_schmidt

DrMZ.ic_error

DrMZ.ifft_norm

DrMZ.inviscid_burgers_pde!

DrMZ.kdv_pde!

DrMZ.load_data

DrMZ.load_data_initial_conditions

DrMZ.load_data_train_test

DrMZ.load_model

DrMZ.loss_all

DrMZ.min_max_scaler

DrMZ.min_max_transform

DrMZ.mse_error

DrMZ.norm_rel_error

DrMZ.opnn_advection_diffusion_pde!

DrMZ.opnn_inviscid_burgers_pde!

DrMZ.opnn_viscous_burgers_pde!

DrMZ.orthonormal_check

DrMZ.periodic_fill_domain

DrMZ.periodic_fill_solution

DrMZ.predict

DrMZ.predict_min_max

DrMZ.quadratic_nonlinear

DrMZ.quadratic_nonlinear_opnn

DrMZ.quadratic_nonlinear_opnn_pseudo

DrMZ.quadratic_nonlinear_pde!

DrMZ.quadratic_nonlinear_triple_product

DrMZ.reduced_initial_condition

DrMZ.save_basis

DrMZ.save_data

DrMZ.save_model

DrMZ.second_derivative_opnn

DrMZ.second_derivative_product

DrMZ.simpson

DrMZ.simpson13

DrMZ.simpson38

DrMZ.solution_extraction

DrMZ.solution_interpolation

DrMZ.solution_spatial_sampling

DrMZ.solution_temporal_sampling

DrMZ.spectral_approximation

DrMZ.spectral_coefficients

DrMZ.spectral_matrix

DrMZ.standard_scaler

DrMZ.standard_transform

DrMZ.train_model

DrMZ.trapz

DrMZ.trapz1

DrMZ.van_leer_limiter

DrMZ.viscous_burgers_pde!

DrMZ.advection_diffusion_pde! — Method

advection_diffusion_pde!(duhat,uhat,p,t_span)

To Do: Overload for real initial condition which only evolves positive modes and complex initial conditions which evolves all modes

DrMZ.advection_pde! — Method

advection_pde!(duhat,uhat,k,t)

RHS for the advection equation $u_t = - u_x$ for numerical integration in Fourier space.

DrMZ.average_error — Method

average_error(error)

DrMZ.average_ic_error — Method

average_ic_error(target,prediction)

DrMZ.backward_upwind — Method

function backward_upwind(u_j,u_jneg,u_jnegg,nu)

DrMZ.backward_upwind_limited — Method

function backward_upwind_limited(u_j,u_jneg,u_jnegg,u_jpos,nu)

DrMZ.basis_OpNN — Method

function basis_OpNN(trunk,x_locations)

DrMZ.build_basis — Method

build_basis(trunk,intial_condition,x_locations,opnn_output_width,initial_condition)

DrMZ.build_basis_redefinition — Method

build_basis_redefinition(basis_full,x_redefined,x_full)

DrMZ.build_dense_model — Method

function build_dense_model(number_layers,neurons,activations)

Build a feedforward neural network (FFNN) consisting of number_layers of Flux dense layers for the specified number of neurons and activations.

Examples

julia> build_dense_model(2,[(128,128),(128,128)],[relu,relu])
Chain(Dense(128, 128, NNlib.relu), Dense(128, 128, NNlib.relu))

DrMZ.cheby_diff — Method

cheby_diff(sol,N,dL)

DrMZ.cheby_diff_matrix — Method

cheby_diff_matrix(N,a,b)

DrMZ.cheby_grid — Method

cheby_grid(N,a,b)

DrMZ.error_test_rel — Method

error_test_rel(target,prediction)

DrMZ.error_test_sse — Method

error_test_sse(target,prediction)

DrMZ.exp_kernel_periodic — Method

function exp_kernel_periodic(x_locations;length_scale=0.5)

Covariance kernel for radial basis function (GRF) and periodic IC $f( sin^2(πx))$.

DrMZ.fft_norm — Method

fft_norm(solution)

DrMZ.forward_upwind — Method

function forward_upwind(u_j,u_jpos,u_jposs,nu)

DrMZ.forward_upwind_limited — Method

function forward_upwind_limited(u_j,u_jpos,u_jposs,u_jneg,nu)

DrMZ.fourier_diff — Method

fourier_diff(sol,N,dL)

DrMZ.generate_basis_solution — Method

generate_basis_solution(L1,L2,t_span,N,basis,initial_condition)

DrMZ.generate_basis_solution_nonlinear — Method

generate_basis_solution_nonlinear(L1,L2,t_span,N,basis,initial_condition)

DrMZ.generate_bwlimitersoupwind_solution — Method

function generate_bwlimitersoupwind_solution(L1,L2,t_end,N,initial_condition;dt=1e-4)

Generate second order limited upwind solution for the inviscid Burgers equation $u_t = -u u_x$ based on the Beam and Warming scheme with a Van Leer limiter.

DrMZ.generate_fourier_solution — Method

DrMZ.generate_fourier_solution_split — Method

generate_fourier_solution_split(L1,L2,tspan,N,initial_condition,n_linear,rhs_sign,nu,pde_function_explicit;dt_fixed=1e-3,rtol=1e-10,atol=1e-14)

n_linear: 2 for viscous Burgers, 3 for KdV; order of linear derivative

rhssign: 1 for viscous Burgers, -1 for KdV; ut = -uux + (rhssign)*(uxx or uxxx)

DrMZ.generate_periodic_functions — Method

generate_periodic_functions(x_locations,number_functions,length_scale)

Generate a specified number_functions of random periodic functions using the exp_kernel_periodic function and a multivariate distribution.

DrMZ.generate_periodic_train_test — Method

generate_periodic_train_test(L1,L2,t_span,number_sensors,number_test_functions,number_train_functions,number_solution _points;length_scale=0.5,batch=number_solution_points,dt=1e-3)

FINISH!!!

DrMZ.generate_sinusoidal_functions_2_parameter — Method

generate_sinusoidal_functions_2_parameter(x_locations,number_functions)

Generate a specified number_functions of random periodic functions for the distribution $α sin(x)+β$ for $α ∈ [-1,1]$ and $β ∈ [-1,1]$

DrMZ.gradient_ratio_backward_j — Method

function gradient_ratio_backward_j(u_j,u_jneg,u_jpos)

DrMZ.gradient_ratio_backward_jneg — Method

function gradient_ratio_backward_jneg(u_j,u_jneg,u_jnegg)

DrMZ.gradient_ratio_forward_j — Method

function gradient_ratio_forward_j(u_j,u_jneg,u_jpos)

DrMZ.gradient_ratio_forward_jpos — Method

function gradient_ratio_forward_jpos(u_j,u_jpos,u_jposs)

DrMZ.gram_schmidt — Method

gram_schmidt(A;sorted="false",tol = 1e-8)

DrMZ.ic_error — Method

ic_error(target,prediction)

DrMZ.ifft_norm — Method

ifft_norm(solution)

DrMZ.inviscid_burgers_pde! — Method

inviscid_burgers_pde!(duhat,uhat,p,t_span)

To Do: Overload for real initial condition which only evolves positive modes and complex initial conditions which evolves all modes

DrMZ.kdv_pde! — Method

kdv_pde!(duhat,uhat,p,t_span)

To Do: Overload for real initial condition which only evolves positive modes and complex initial conditions which evolves all modes

DrMZ.load_data — Method

load_data(n_epoch,number_train_functions,number_test_functions,pde_function)

Load the trained branch and trunk neural networks along with the train_data and test_data.

DrMZ.load_data_initial_conditions — Method

load_data_initial_conditions(number_train_functions,number_test_functions)

FINISH!!!

DrMZ.load_data_train_test — Method

load_data_train_test(number_train_functions,number_test_functions,pde_function)

Load the train_data and test_data.

DrMZ.load_model — Method

load_model(n_epoch,pde_function)

Load the trained branch and trunk neural networks.

DrMZ.loss_all — Method

loss_all(branch,trunk,initial_conditon,solution_location,target_value)

Compute the mean squared error (MSE) for a complete dataset.

DrMZ.min_max_scaler — Method

min_max_scaler(x;dim = 2)

DrMZ.min_max_transform — Method

min_max_transform(x,scale_object;min = 0,max = 1)

DrMZ.mse_error — Method

mse_error(target,prediction)

DrMZ.norm_rel_error — Method

norm_rel_error(target,prediction)

DrMZ.opnn_advection_diffusion_pde! — Method

opnn_advection_diffusion_pde!(du,u,p,t_span)

DrMZ.opnn_inviscid_burgers_pde! — Method

opnn_inviscid_burgers_pde!(du,u,p,t_span)

DrMZ.opnn_viscous_burgers_pde! — Method

opnn_viscous_burgers_pde!(du,u,p,t_span)

DrMZ.orthonormal_check — Method

orthonormal_check(basis;tol = 1e-15)

DrMZ.periodic_fill_domain — Method

periodic_fill_domain(x_locations)

DrMZ.periodic_fill_solution — Method

periodic_fill_solution(solution)

DrMZ.predict — Method

predict(branch,trunk,initial_condition,x_locations,t_values)

Predict solution $u(t,x)$ at specified output locations using trained operator neural network branch and trunk.

DrMZ.predict_min_max — Method

predict_min_max(branch,trunk,initial_condition,x_locations,t_values,scale_object)

Uses the trained operator neural net branch and trunk to predict solution at specified output locations when using min-max normalization

DrMZ.quadratic_nonlinear — Method

quadratic_nonlinear!(uhat)

To Do: Overload for real initial condition which only evolves positive modes and complex initial conditions which evolves all modes

DrMZ.quadratic_nonlinear_opnn — Method

quadratic_nonlinear_opnn(uhat)

DrMZ.quadratic_nonlinear_opnn_pseudo — Method

quadratic_nonlinear_opnn_pseudo(basis,uhat)

DrMZ.quadratic_nonlinear_pde! — Method

quadratic_nonlinear_pde!(duhat,uhat,p,t_span)

To Do: Overload for real initial condition which only evolves positive modes and complex initial conditions which evolves all modes

DrMZ.quadratic_nonlinear_triple_product — Method

quadratic_nonlinear_triple_product(basis,Dbasis)

DrMZ.reduced_initial_condition — Method

reduced_initial_condition(L1,L2,N,x_locations,initial_condition)

DrMZ.save_basis — Method

save_basis(basis,N,n_epoch,pde_function)

FINISH!!!

DrMZ.save_data — Method

save_data(train_data,test_data,number_train_functions,number_test_functions,number_solution_points,pde_function)

Save the train_data and test_data.

DrMZ.save_model — Method

save_model(branch,trunk,n_epoch,loss_all_train,loss_all_test,pde_function)

Save the trained branch and trunk neural networks and the training and testing loss history.

DrMZ.second_derivative_opnn — Method

second_derivative_opnn(uhat)

DrMZ.second_derivative_product — Method

second_derivative_product(basis,DDbasis)

DrMZ.simpson — Method

simpson(x_range,integrand)

Numerical integration using the multi-application Simpson's and trapezoid rules.

DrMZ.simpson13 — Method

simpson13(h,f1,f2)

Numerical integration using the single-application Simpson's 1/3 rule.

DrMZ.simpson38 — Method

simpson38(x_range,integrand)

Numerical integration using the single-application Simpson's 3/8 rule.

DrMZ.solution_extraction — Method

solution_extraction(x_locations,t_values,solution,initial_condition,number_solution_points)

Extract the specified number_solution_points randomly from the u(t,x) solution space.

DrMZ.solution_interpolation — Method

solution_interpolation(t_span_original,x_locations_original,t_span_interpolate,x_locations_interpolate,solution)

DrMZ.solution_spatial_sampling — Method

solution_spatial_sampling(x_prediction,x_target,solution)

DrMZ.solution_temporal_sampling — Method

solution_temporal_sampling(t_prediction,t_target,solution)

DrMZ.spectral_approximation — Method

spectral_approximation(basis,coefficients)

DrMZ.spectral_coefficients — Method

spectral_coefficients(basis,fnc)

DrMZ.spectral_matrix — Method

spectral_matrix(basis,Dbasis)

DrMZ.standard_scaler — Method

standard_scaler(x;dim=2)

DrMZ.standard_transform — Method

standard_transform(x,scale_object)

DrMZ.train_model — Method

train_model(branch,trunk,n_epoch,train_data;learning_rate=0.00001)

Train the operator neural network using the mean squared error (MSE) and ADAM optimization for n_epochs epochs.

DrMZ.trapz — Method

trapz(x_range,integrand)

Numerical integration using the multi-application trapezoidal rule.

DrMZ.trapz1 — Method

trapz1(x_range,integrand)

Numerical integration using the single-application trapezoidal rule.

DrMZ.van_leer_limiter — Method

function van_leer_limiter(r)

DrMZ.viscous_burgers_pde! — Method

viscous_burgers_pde!(duhat,uhat,p,t_span)

To Do: Overload for real initial condition which only evolves positive modes and complex initial conditions which evolves all modes