DrMZ

DrMZ

DrMZ.advection_diffusion_pde! — Method

advection_diffusion_pde!(duhat,uhat,p,t)

RHS for the advection-diffusion equation $u_t = - u_x + ν u_{xx}$ for numerical integration in Fourier space where ν is the viscosity.

DrMZ.advection_pde! — Method

advection_pde!(duhat,uhat,p,t)

RHS for the advection equation $u_t = - u_x$ for numerical integration in Fourier space.

DrMZ.average_error — Method

average_error(domain,error)

Compute the average error using the trapezoid rule $\frac{1}{T} \int_0^T error(t) dt$.

DrMZ.average_ic_error — Method

average_ic_error(target,prediction)

Compute the two-norm relative error between the prediction and target values for an initial condition.

DrMZ.backward_upwind — Method

function backward_upwind(u_j,u_jneg,u_jnegg,nu)

Compute the partial RHS ($u_j^{n+1} = u_j^n$ + backward_upwind) of the second order backward upwind scheme of Beam and Warming for the inviscid Burgers equation $u_t = -u u_x$ (split into monotone and non-montone terms), $-\nu(u_j^n-u_{j-1}^n) - \frac{\nu}{2}(1-\nu)(u_j^n-u_{j-1}^n) + \frac{\nu}{2}(1-\nu)(u_{j-1}^n-u_{j-2}^n)$. nu is the CFL condition.

DrMZ.backward_upwind_limited — Method

function backward_upwind_limited(u_j,u_jneg,u_jnegg,u_jpos,nu)

Compute the partial RHS ($u_j^{n+1} = u_j^n$ + backward_upwind) of the second order limited backward upwind scheme of Beam and Warming for the inviscid Burgers equation $u_t = -u u_x$ (split into monotone and non-montone terms), $-\nu(u_j^n-u_{j-1}^n) - \frac{\nu}{2}(1-\nu)(u_j^n-u_{j-1}^n) + \frac{\nu}{2}(1-\nu)(u_{j-1}^n-u_{j-2}^n)$. nu is the CFL condition.

DrMZ.basis_OpNN — Method

function basis_OpNN(trunk,x_locations)

DrMZ.build_basis — Method

build_basis(trunk,intial_condition,x_locations,opnn_output_width,initial_condition)

DrMZ.build_basis_redefinition — Method

build_basis_redefinition(basis_full,x_redefined,x_full)

DrMZ.build_dense_model — Method

function build_dense_model(number_layers,neurons,activations)

Build a feedforward neural network (FFNN) consisting of number_layers of Flux dense layers for the specified number of neurons and activations.

Examples

julia> build_dense_model(2,[(128,128),(128,128)],[relu,relu])
Chain(Dense(128, 128, NNlib.relu), Dense(128, 128, NNlib.relu))

DrMZ.central_difference — Method

function central_difference(u_j,u_jpos,u_jneg,mu)

Compute the second order central difference for the viscous term of the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$. mu is equal to $ \frac{\nu \Delta t}{\Delta x^2}$.

DrMZ.cheby_diff — Method

cheby_diff(sol,N,dL)

Compute the derivative using a Chebyshev differentiation matrix on the interval $[a,b]$ with N discretization points.

DrMZ.cheby_diff_matrix — Method

cheby_diff_matrix(N,a,b)

Generate the Chebyshev differentiation matrix for the interval $[a,b]$ with N discretization points.

DrMZ.cheby_grid — Method

cheby_grid(N,a,b)

Generate the grid of Chebyshev points on the interval $[a,b]$ with N discretization points.

DrMZ.error_rel — Method

error_rel(target,prediction)

Compute the relative error between the prediction and target values.

DrMZ.error_se — Method

error_se(target,prediction)

Compute the squared error between the prediction and target values.

DrMZ.exp_kernel_periodic — Method

function exp_kernel_periodic(x_locations;length_scale=0.5)

Covariance kernel for radial basis function (GRF) and periodic IC $f(\sin^2(πx))$.

DrMZ.fft_norm — Method

fft_norm(solution)

Compute the FFT normalized by $\frac{1}{N}$.

DrMZ.forward_upwind — Method

function forward_upwind(u_j,u_jpos,u_jposs,nu)

Compute the partial RHS ($u_j^{n+1} = u_j^n$ + forward_upwind) of the second order forward upwind scheme of Beam and Warming for the inviscid Burgers equation $u_t = -u u_x$ (split into monotone and non-montone terms), $-\nu(u_{j+1}^n-u_{j}^n) - \frac{\nu}{2}(\nu+1)(u_{j+1}^n-u_{j}^n) + \frac{\nu}{2}(\nu+1)(u_{j+2}^n-u_{j+1}^n)$. nu is the CFL condition.

DrMZ.forward_upwind_limited — Method

function forward_upwind_limited(u_j,u_jpos,u_jposs,u_jneg,nu)

Compute the partial RHS ($u_j^{n+1} = u_j^n$ + forward_upwind) of the second order limited forward upwind scheme of Beam and Warming for the inviscid Burgers equation $u_t = -u u_x$ (split into monotone and non-montone terms), $-\nu(u_{j+1}^n-u_{j}^n) - \frac{\nu}{2}(\nu+1)(u_{j+1}^n-u_{j}^n) + \frac{\nu}{2}(\nu+1)(u_{j+2}^n-u_{j+1}^n)$. nu is the CFL condition.

DrMZ.fourier_diff — Method

fourier_diff(sol,N,dL;format="matrix")

Compute the derivative using a Fourier differentiation matrix (default) or the spectral derivative for periodic functions for domain length dL and with N discretization points.

DrMZ.gauss_quad — Method

function gauss_quad(a,b,func,number_points)

Compute the integral using Gauss-Legendre quadrature for the interval $\int_a^b$ for a given func using a specified number_points.

DrMZ.generate_basis_solution — Method

generate_basis_solution(L1,L2,t_span,N,basis,initial_condition,pde_function;dt=1e-3,nu=0.1,rtol=1e-10,atol=1e-14)

Generate the solution for a given linear pde_function and initial_conditon on a periodic domain using a N mode custom basis expansion.

DrMZ.generate_basis_solution_nonlinear — Method

generate_basis_solution_nonlinear(L1,L2,t_span,N,basis,initial_condition,pde_function;dt=1e-4,nu=0.1,rtol=1e-10,atol=1e-14)

Generate the solution for a given non-linear pde_function and initial_conditon on a periodic domain using a N mode custom basis expansion.

DrMZ.generate_bwlimitersoupwind_solution — Method

function generate_bwlimitersoupwind_solution(L1,L2,t_end,N,initial_condition;dt=1e-4)

Generate second order limited upwind solution for the inviscid Burgers equation $u_t = -u u_x$ based on the Beam and Warming scheme with a Van Leer limiter.

DrMZ.generate_bwlimitersoupwind_viscous_solution — Method

function generate_bwlimitersoupwind_viscous_solution(L1,L2,t_end,N,initial_condition;dt=1e-4,nu=0.1)

Generate second order solution for the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$ based on the Beam and Warming second order upwind with a Van Leer limiter for the convective term and a second order central differnce for the diffusive term. nu is the viscosity.

DrMZ.generate_fourier_solution — Method

generate_fourier_solution(L1,L2,tspan,N,initial_condition,pde_function;dt=1e-3,nu=0.1,rtol=1e-10,atol=1e-14)

Generate the solution for a given pde_function and initial_condition on a periodic domain using a N mode Fourier expansion.

DrMZ.generate_periodic_functions — Method

generate_periodic_functions(x_locations,number_functions,length_scale)

Generate a specified number_functions of random periodic vectors using the exp_kernel_periodic function and a multivariate distribution.

DrMZ.generate_periodic_train_test — Method

generate_periodic_train_test(L1,L2,t_span,number_sensors,number_test_functions,number_train_functions,number_solution_points,pde_function_handle;length_scale=0.5,batch=number_solution_points,dt=1e-3,domain="periodic"))

Generate the training and testing data for a specified pde_function_handle for periodic boundary conditions using a Fourier spectral method.

DrMZ.generate_sinusoidal_functions_2_parameter — Method

generate_sinusoidal_functions_2_parameter(x_locations,number_functions)

Generate a specified number_functions of random periodic vectors for the distribution $α \sin(x)+β$ for $α ∈ [-1,1]$ and $β ∈ [-1,1]$.

DrMZ.get_1D_energy_basis — Method

function get_1D_energy_basis(basis,u_solution)

Compute the energy in the custom basis space. The expansion coefficients are scaled by $ \frac{1}{\sqrt(N)} $ for comparison to the non-unitary Fourier coefficients.

DrMZ.get_1D_energy_fft — Method

function get_1D_energy_fft(u_solution)

Compute the energy in the Fourier domain using the scaling of $ \frac{1}{N} $.

DrMZ.gradient_ratio_backward_j — Method

function gradient_ratio_backward_j(u_j,u_jneg,u_jpos)

Compute the backward gradient ratio at point $i$ as required for the van_leer_limiter.

DrMZ.gradient_ratio_backward_jneg — Method

function gradient_ratio_backward_jneg(u_j,u_jneg,u_jnegg)

Compute the backward gradient ratio at point $i-1$ as required for the van_leer_limiter.

DrMZ.gradient_ratio_forward_j — Method

function gradient_ratio_forward_j(u_j,u_jneg,u_jpos)

Compute the forward gradient ratio at point $i$ as required for the van_leer_limiter.

DrMZ.gradient_ratio_forward_jpos — Method

function gradient_ratio_forward_jpos(u_j,u_jpos,u_jposs)

Compute the forward gradient ratio at point $i+1$ as required for the van_leer_limiter.

DrMZ.ic_error — Method

ic_error(target,prediction)

Compute the relative error between the prediction and target values for an initial condition. If the target = 0, both the prediction and target are augmented by $\epsilon_{machine}$.

DrMZ.ifft_norm — Method

ifft_norm(solution)

Compute the IFFT normalized by $N$.

DrMZ.inviscid_burgers_pde! — Method

inviscid_burgers_pde!(duhat,uhat,p,t)

RHS for the inviscid Burgers equation $u_t = - u u_x$ for numerical integration in Fourier space.

DrMZ.load_data — Method

load_data(n_epoch,number_train_functions,number_test_functions,pde_function)

Load the trained branch and trunk neural networks along with the train_data and test_data.

DrMZ.load_data_initial_conditions — Method

load_data_initial_conditions(number_train_functions,number_test_functions)

Load the initial conditions from the train_data and test_data.

DrMZ.load_data_train_test — Method

load_data_train_test(number_train_functions,number_test_functions,pde_function)

Load the train_data and test_data.

DrMZ.load_model — Method

load_model(n_epoch,pde_function)

Load the trained branch and trunk neural networks.

DrMZ.loss_all — Method

loss_all(branch,trunk,initial_conditon,solution_location,target_value)

Compute the mean squared error (MSE) for a complete dataset.

DrMZ.min_max_scaler — Method

min_max_scaler(x;dim = 2)

Compute the required scaler to shift the features to a given range.

DrMZ.min_max_transform — Method

min_max_transform(x,scale_object;min = 0,max = 1)

Apply min_max_scaler to shift the features.

DrMZ.mse_error — Method

mse_error(target,prediction)

Compute the mean squared error between the prediction and target values.

DrMZ.norm_rel_error — Method

norm_rel_error(target,prediction)

Compute the two-norm relative error between the prediction and target values.

DrMZ.opnn_advection_diffusion_pde! — Method

opnn_advection_diffusion_pde!(du,u,p,t)

RHS for the advection-diffusion equation $u_t = - u_x + ν u_{xx}$ for numerical integration in the custom basis space where ν is the viscosity.

DrMZ.opnn_advection_pde! — Method

opnn_advection_pde!(du,u,p,t)

RHS for the advection equation $u_t = - u_x$ for numerical integration in the custom basis space.

DrMZ.opnn_inviscid_burgers_pde! — Method

opnn_inviscid_burgers_pde!(du,u,p,t)

RHS for the inviscid Burgers equation $u_t = - u u_x$ for numerical integration in the custom basis space.

DrMZ.opnn_viscous_burgers_pde! — Method

opnn_viscous_burgers_pde!(du,u,p,t)

RHS for the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$ for numerical integration in the custom basis space where ν is the viscosity.

DrMZ.orthonormal_check — Method

orthonormal_check(basis;tol = 1e-15)

DrMZ.periodic_fill_domain — Method

periodic_fill_domain(x_locations)

Output the full domain from periodic domain specified for x_locations.

DrMZ.periodic_fill_solution — Method

periodic_fill_solution(solution)

Output the full $u(t,x)$ solution from periodic solution.

DrMZ.predict — Method

predict(branch,trunk,initial_condition,x_locations,t_values)

Predict solution $u(t,x)$ at specified output locations using trained operator neural network branch and trunk.

DrMZ.predict_min_max — Method

predict_min_max(branch,trunk,initial_condition,x_locations,t_values,scale_object)

Uses the trained operator neural net branch and trunk to predict solution at specified output locations when using min-max normalization

DrMZ.quadratic_nonlinear — Method

quadratic_nonlinear!(uhat,N,dL,alpha)

Compute the convolution sum $\frac{ik}{2}\sum_{p+q=k} u_p u_q$ resulting from the quadratic nonlinearity of Burgers equation $u u_x$ in Fourier space. Convolution sum is padded with the 3/2 rule for dealiasing.

DrMZ.quadratic_nonlinear_opnn — Method

quadratic_nonlinear_opnn(uhat,nonlinear_triple)

Compute the triple product sum $\sum_{k=1}^N \sum_{l=1}^N u_k u_l \sum_{j=0}^{N-1} \phi_{jk} \phi_{jl}^{'} \phi_{jm}^{*}$ resulting from the quadratic nonlinearity of Burgers equation $u u_x$ in custom basis space.

DrMZ.quadratic_nonlinear_opnn_pseudo — Method

quadratic_nonlinear_opnn_pseudo(basis,uhat,dL,N)

Compute the quadratic nonlinearity of Burgers equation $u u_x$ in custom basis space using a pseudo spectral type approach. At each step, transform solution to real space and compute the product of the real space solution and the spatial derivative of the real space solution prior to transforming back to custom basis space.

DrMZ.quadratic_nonlinear_triple_product — Method

quadratic_nonlinear_triple_product(basis,Dbasis)

DrMZ.reduced_initial_condition — Method

reduced_initial_condition(L1,L2,N,x_locations,initial_condition)

Extract the $x$ locations and intial condition values $u(x)$ at a reduced number of equally spaced spatial locations.

DrMZ.save_basis — Method

save_basis(basis,N,n_epoch,pde_function)

DrMZ.save_data — Method

save_data(train_data,test_data,number_train_functions,number_test_functions,number_solution_points,pde_function)

Save the train_data and test_data.

DrMZ.shifted_nodes — Method

function shifted_nodes(a,b,xd)

Compute the shifted nodes for Gauss-Legendre quadrature from $\int_{-1}^1$ to $ \int_a^b$.

DrMZ.simpson — Method

simpson(x_range,integrand)

Numerical integration using the multi-application Simpson's and trapezoid rules.

DrMZ.simpson13 — Method

simpson13(h,f1,f2)

Numerical integration using the single-application Simpson's 1/3 rule.

DrMZ.simpson38 — Method

simpson38(x_range,integrand)

Numerical integration using the single-application Simpson's 3/8 rule.

DrMZ.solution_extraction — Method

solution_extraction(x_locations,t_values,solution,initial_condition,number_solution_points)

Extract the specified number_solution_points randomly from the $u(t,x)$ solution space.

DrMZ.solution_interpolation — Method

solution_interpolation(t_span_original,x_locations_original,t_span_interpolate,x_locations_interpolate,solution)

Compute the $u(t,x)$ solution at intermediate $(t,x)$ locations using linear interpolation.

DrMZ.solution_spatial_sampling — Method

solution_spatial_sampling(x_prediction,x_target,solution)

Extract the solution values $u(t,x)$ at a reduced number of equally spaced spatial locations.

DrMZ.solution_temporal_sampling — Method

solution_temporal_sampling(t_prediction,t_target,solution)

Extract the solution values $u(t,x)$ at a reduced number of equally spaced temporal locations.

DrMZ.spectral_approximation — Method

spectral_approximation(basis,coefficients)

DrMZ.spectral_coefficients — Method

spectral_coefficients(basis,fnc)

DrMZ.spectral_matrix — Method

spectral_matrix(basis,Dbasis)

DrMZ.standard_scaler — Method

standard_scaler(x;dim=2)

Compute the required scaler to shift the features to have zero mean and unit variance.

DrMZ.standard_transform — Method

standard_transform(x,scale_object)

Apply standard_scaler to shift the features.

DrMZ.train_model — Method

train_model(branch,trunk,n_epoch,train_data;learning_rate=0.00001,save_at=2500)

Train the operator neural network using the mean squared error (MSE) and Adam optimization for n_epochs epochs.

DrMZ.trapz — Method

trapz(x_range,integrand)

Numerical integration using the multi-application trapezoidal rule.

DrMZ.trapz1 — Method

trapz1(x_range,integrand)

Numerical integration using the single-application trapezoidal rule.

DrMZ.van_leer_limiter — Method

function van_leer_limiter(r)

Compute the Van Leer limiter $\Psi = \frac{r + |r|}{1+r}$ where $r$ is the gradient ratio.

DrMZ.viscous_burgers_pde! — Method

viscous_burgers_pde!(duhat,uhat,p,t)

RHS for the viscous Burgers equation $u_t = - u u_x + ν u_{xx}$ for numerical integration in Fourier space where ν is the viscosity.

DrMZ.advection_diffusion_pde!

DrMZ.advection_pde!

DrMZ.average_error

DrMZ.average_ic_error

DrMZ.backward_upwind

DrMZ.backward_upwind_limited

DrMZ.basis_OpNN

DrMZ.build_basis

DrMZ.build_basis_redefinition

DrMZ.build_dense_model

DrMZ.central_difference

DrMZ.cheby_diff

DrMZ.cheby_diff_matrix

DrMZ.cheby_grid

DrMZ.error_rel

DrMZ.error_se

DrMZ.exp_kernel_periodic

DrMZ.fft_norm

DrMZ.forward_upwind

DrMZ.forward_upwind_limited

DrMZ.fourier_diff

DrMZ.gauss_quad

DrMZ.generate_basis_solution

DrMZ.generate_basis_solution_nonlinear

DrMZ.generate_bwlimitersoupwind_solution

DrMZ.generate_bwlimitersoupwind_viscous_solution

DrMZ.generate_fourier_solution

DrMZ.generate_periodic_functions

DrMZ.generate_periodic_train_test

DrMZ.generate_sinusoidal_functions_2_parameter

DrMZ.get_1D_energy_basis

DrMZ.get_1D_energy_fft

DrMZ.gradient_ratio_backward_j

DrMZ.gradient_ratio_backward_jneg

DrMZ.gradient_ratio_forward_j

DrMZ.gradient_ratio_forward_jpos

DrMZ.ic_error

DrMZ.ifft_norm

DrMZ.inviscid_burgers_pde!

DrMZ.load_data

DrMZ.load_data_initial_conditions

DrMZ.load_data_train_test

DrMZ.load_model

DrMZ.loss_all

DrMZ.min_max_scaler

DrMZ.min_max_transform

DrMZ.mse_error

DrMZ.norm_rel_error

DrMZ.opnn_advection_diffusion_pde!

DrMZ.opnn_advection_pde!

DrMZ.opnn_inviscid_burgers_pde!

DrMZ.opnn_viscous_burgers_pde!

DrMZ.orthonormal_check

DrMZ.periodic_fill_domain

DrMZ.periodic_fill_solution

DrMZ.predict

DrMZ.predict_min_max

DrMZ.quadratic_nonlinear

DrMZ.quadratic_nonlinear_opnn

DrMZ.quadratic_nonlinear_opnn_pseudo

DrMZ.quadratic_nonlinear_triple_product

DrMZ.reduced_initial_condition

DrMZ.save_basis

DrMZ.save_data

DrMZ.shifted_nodes

DrMZ.simpson

DrMZ.simpson13

DrMZ.simpson38

DrMZ.solution_extraction

DrMZ.solution_interpolation

DrMZ.solution_spatial_sampling

DrMZ.solution_temporal_sampling

DrMZ.spectral_approximation

DrMZ.spectral_coefficients

DrMZ.spectral_matrix

DrMZ.standard_scaler

DrMZ.standard_transform

DrMZ.train_model

DrMZ.trapz

DrMZ.trapz1

DrMZ.van_leer_limiter

DrMZ.viscous_burgers_pde!